kata(arrays): Twin Strings
It was a bit challenging, for me.

I wasn't able to finish this exercise during the 30 mins of a job test. The reason was that I couldn't understand the exercise enunciation at the time and I didn't know the Ruby API for even/odd numbers.

In fact, I couldn't even compile my code. Shame! Shame! 😅

Afterwards, at home, I managed to finish the test, which improved my skills with simple evens/odds problems.